Refactor to move files to the correct folder (outside public)
