+++ src/main/routes/features/claim/claimantPartyTypeController.ts
@@ -5,7 +5,7 @@ import {PartyTypeSelection} from '../../../common/form/models/claim/partyTypeSel
 import {ClaimantOrDefendant} from '../../../common/models/partyType';
 import {redirectToPage} from '../../../services/features/claim/partyTypeService';
 
-const claimantPartyTypeViewPath = 'features/public/claim/claimant-party-type';
+const claimantPartyTypeViewPath = 'features/claim/claimant-party-type';
 const claimantPartyTypeController = express.Router();
 
 claimantPartyTypeController.get(CLAIMANT_PARTY_TYPE_SELECTION_URL, (req: express.Request, res: express.Response) => {
