+++ src/main/views/features/claim/claimant-party-type.njk
@@ -1,10 +1,10 @@
 {% extends "claim-details-tpl.njk" %}
 {% from "govuk/components/button/macro.njk" import govukButton %}
 {% from "govuk/components/radios/macro.njk" import govukRadios %}
-{% from "../../../macro/contact-us-for-help.njk" import contactUsForHelp %}
-{% from "../../../macro/csrf.njk" import csrfProtection %}
-{% from "../../../macro/page-title.njk" import setPageTitle %}
-{% from "../../../macro/error-summary.njk" import errorSummary %}
+{% from "../../macro/contact-us-for-help.njk" import contactUsForHelp %}
+{% from "../../macro/csrf.njk" import csrfProtection %}
+{% from "../../macro/page-title.njk" import setPageTitle %}
+{% from "../../macro/error-summary.njk" import errorSummary %}
 {% block pageTitle %}
   {{ setPageTitle('PAGES.CLAIMANT_PARTY_TYPE_SELECTION.PAGE_TITLE') }}
 {% endblock %}
