+++ src/test/unit/routes/features/claim/claimantDoBController.test.ts
@@ -1,13 +1,13 @@
 import config from 'config';
 import nock from 'nock';
 import request from 'supertest';
-import {app} from '../../../../../../main/app';
-import {CLAIMANT_DOB_URL, CLAIMANT_PHONE_NUMBER_URL} from '../../../../../../main/routes/urls';
-import {mockCivilClaim, mockNoStatementOfMeans, mockRedisFailure} from '../../../../../utils/mockDraftStore';
-import {TestMessages} from '../../../../../utils/errorMessageTestConstants';
+import {app} from '../../../../../main/app';
+import {CLAIMANT_DOB_URL, CLAIMANT_PHONE_NUMBER_URL} from '../../../../../main/routes/urls';
+import {mockCivilClaim, mockNoStatementOfMeans, mockRedisFailure} from '../../../../utils/mockDraftStore';
+import {TestMessages} from '../../../../utils/errorMessageTestConstants';
 
-jest.mock('../../../../../../main/modules/oidc');
-jest.mock('../../../../../../main/modules/draft-store');
+jest.mock('../../../../../main/modules/oidc');
+jest.mock('../../../../../main/modules/draft-store');
 
 describe('Claimant Date of Birth Controller', () => {
   const citizenRoleToken: string = config.get('citizenRoleToken');
