+++ src/test/unit/views/features/claim/defendant-party-type.test.ts
@@ -1,14 +1,14 @@
 import config from 'config';
 import nock from 'nock';
 import request from 'supertest';
-import {app} from '../../../../../../main/app';
-import {mockCivilClaim} from '../../../../../utils/mockDraftStore';
-import {CLAIM_DEFENDANT_PARTY_TYPE_URL} from '../../../../../../main/routes/urls';
+import {app} from '../../../../../main/app';
+import {mockCivilClaim} from '../../../../utils/mockDraftStore';
+import {CLAIM_DEFENDANT_PARTY_TYPE_URL} from '../../../../../main/routes/urls';
 
 const jsdom = require('jsdom');
 const {JSDOM} = jsdom;
-jest.mock('../../../../../../main/modules/oidc');
-jest.mock('../../../../../../main/modules/draft-store');
+jest.mock('../../../../../main/modules/oidc');
+jest.mock('../../../../../main/modules/draft-store');
 
 describe('Defendant Party View', () => {
   const citizenRoleToken: string = config.get('citizenRoleToken');
